refactor(docs): move test-only modules out of apps/docs/src/lib (#992)
lib/ should mean what the site ships. color-evidence and gallery-filter-journey
were only imported by tests; playground-samples only fed the seed generator.

Move them beside their owners and update the four importers. No behaviour change.
