+++ scripts/color-evidence.ts
